+++ exporters/otlp/src/otlp_recordable.cc
@@ -42,16 +42,16 @@ namespace otlp
 void OtlpRecordable::SetIdentity(const opentelemetry::trace::SpanContext &span_context,
                                  opentelemetry::trace::SpanId parent_span_id) noexcept
 {
-  span_.set_trace_id(reinterpret_cast<const char *>(span_context.trace_id().Id().data()),
-                     trace::TraceId::kSize);
-  span_.set_span_id(reinterpret_cast<const char *>(span_context.span_id().Id().data()),
-                    trace::SpanId::kSize);
+  span_->set_trace_id(reinterpret_cast<const char *>(span_context.trace_id().Id().data()),
+                      trace::TraceId::kSize);
+  span_->set_span_id(reinterpret_cast<const char *>(span_context.span_id().Id().data()),
+                     trace::SpanId::kSize);
   if (parent_span_id.IsValid())
   {
-    span_.set_parent_span_id(reinterpret_cast<const char *>(parent_span_id.Id().data()),
-                             trace::SpanId::kSize);
+    span_->set_parent_span_id(reinterpret_cast<const char *>(parent_span_id.Id().data()),
+                              trace::SpanId::kSize);
   }
-  span_.set_trace_state(span_context.trace_state()->ToHeader());
+  span_->set_trace_state(span_context.trace_state()->ToHeader());
 }
 
 proto::resource::v1::Resource OtlpRecordable::ProtoResource() const noexcept
@@ -139,13 +139,13 @@ void OtlpRecordable::SetAttribute(nostd::string_view key,
     return;
   }
 
-  if (static_cast<std::uint32_t>(span_.attributes_size()) >= span_limits_.attribute_count_limit)
+  if (static_cast<std::uint32_t>(span_->attributes_size()) >= span_limits_.attribute_count_limit)
   {
-    span_.set_dropped_attributes_count(span_.dropped_attributes_count() + 1);
+    span_->set_dropped_attributes_count(span_->dropped_attributes_count() + 1);
     return;
   }
 
-  auto *attribute = span_.add_attributes();
+  auto *attribute = span_->add_attributes();
   OtlpPopulateAttributeUtils::PopulateAttribute(
       attribute, key, value, AttributeConverterOptions{span_limits_.attribute_value_length_limit});
 }
@@ -154,13 +154,13 @@ void OtlpRecordable::AddEvent(nostd::string_view name,
                               common::SystemTimestamp timestamp,
                               const common::KeyValueIterable &attributes) noexcept
 {
-  if (static_cast<std::uint32_t>(span_.events_size()) >= span_limits_.event_count_limit)
+  if (static_cast<std::uint32_t>(span_->events_size()) >= span_limits_.event_count_limit)
   {
-    span_.set_dropped_events_count(span_.dropped_events_count() + 1);
+    span_->set_dropped_events_count(span_->dropped_events_count() + 1);
     return;
   }
 
-  auto *event = span_.add_events();
+  auto *event = span_->add_events();
   event->set_name(name.data(), name.size());
   event->set_time_unix_nano(timestamp.time_since_epoch().count());
 
@@ -181,12 +181,12 @@ void OtlpRecordable::AddEvent(nostd::string_view name,
 void OtlpRecordable::AddLink(const trace::SpanContext &span_context,
                              const common::KeyValueIterable &attributes) noexcept
 {
-  if (static_cast<std::uint32_t>(span_.links_size()) >= span_limits_.link_count_limit)
+  if (static_cast<std::uint32_t>(span_->links_size()) >= span_limits_.link_count_limit)
   {
-    span_.set_dropped_links_count(span_.dropped_links_count() + 1);
+    span_->set_dropped_links_count(span_->dropped_links_count() + 1);
     return;
   }
-  auto *link = span_.add_links();
+  auto *link = span_->add_links();
   link->set_trace_id(reinterpret_cast<const char *>(span_context.trace_id().Id().data()),
                      trace::TraceId::kSize);
   link->set_span_id(reinterpret_cast<const char *>(span_context.span_id().Id().data()),
@@ -208,23 +208,23 @@ void OtlpRecordable::AddLink(const trace::SpanContext &span_context,
 
 void OtlpRecordable::SetStatus(trace::StatusCode code, nostd::string_view description) noexcept
 {
-  span_.mutable_status()->set_code(proto::trace::v1::Status_StatusCode(code));
+  span_->mutable_status()->set_code(proto::trace::v1::Status_StatusCode(code));
   if (code == trace::StatusCode::kError)
   {
-    span_.mutable_status()->set_message(description.data(), description.size());
+    span_->mutable_status()->set_message(description.data(), description.size());
   }
 }
 
 void OtlpRecordable::SetName(nostd::string_view name) noexcept
 {
-  span_.set_name(name.data(), name.size());
+  span_->set_name(name.data(), name.size());
 }
 
 void OtlpRecordable::SetTraceFlags(opentelemetry::trace::TraceFlags flags) noexcept
 {
   uint32_t all_flags = flags.flags() & opentelemetry::proto::trace::v1::SPAN_FLAGS_TRACE_FLAGS_MASK;
 
-  span_.set_flags(all_flags);
+  span_->set_flags(all_flags);
 }
 
 void OtlpRecordable::SetSpanKind(trace::SpanKind span_kind) noexcept
@@ -260,18 +260,18 @@ void OtlpRecordable::SetSpanKind(trace::SpanKind span_kind) noexcept
       proto_span_kind = proto::trace::v1::Span_SpanKind::Span_SpanKind_SPAN_KIND_UNSPECIFIED;
   }
 
-  span_.set_kind(proto_span_kind);
+  span_->set_kind(proto_span_kind);
 }
 
 void OtlpRecordable::SetStartTime(common::SystemTimestamp start_time) noexcept
 {
-  span_.set_start_time_unix_nano(start_time.time_since_epoch().count());
+  span_->set_start_time_unix_nano(start_time.time_since_epoch().count());
 }
 
 void OtlpRecordable::SetDuration(std::chrono::nanoseconds duration) noexcept
 {
-  const uint64_t unix_end_time = span_.start_time_unix_nano() + duration.count();
-  span_.set_end_time_unix_nano(unix_end_time);
+  const uint64_t unix_end_time = span_->start_time_unix_nano() + duration.count();
+  span_->set_end_time_unix_nano(unix_end_time);
 }
 
 void OtlpRecordable::SetInstrumentationScope(